backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.6
redis==5.0.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
Pillow==10.2.0
django-celery-beat==2.6.0
pywebpush==1.14.0
pytz==2024.1
